image_admin_panel/requirements.txt
flask==2.0.1
openai==0.27.0
pandas==1.3.3
python-dotenv==0.19.0
Pillow==8.3.2
openpyxl==3.0.9
requests==2.26.0 